+++ database/migrations/2021_01_02_000010_create_targets_table.php
@@ -13,9 +13,9 @@ class CreateTargetsTable extends Migration
             $table->string('target_type')->nullable();
             $table->string('target_name');
             $table->string('target_status');
-            $table->decimal('amount', 15, 2);
+            $table->integer('amount');
             $table->date('first_pay_date')->nullable();
-            $table->decimal('monthly_pay_amount', 15, 2)->nullable();
+            $table->integer('monthly_pay_amount')->nullable();
             $table->date('pay_to_date')->nullable();
             $table->longText('description')->nullable();
             $table->boolean('target_is_done')->default(0)->nullable();
